Move gen code into go pkg folder
- Rename ECS to ECSX for consistency.
